+++ selfdrive/car/car_helpers.py
@@ -3,6 +3,7 @@ import time
 from collections.abc import Callable
 
 from cereal import car
+from openpilot.common.numpy_fast import interp
 from openpilot.common.params import Params
 from openpilot.selfdrive.car.interfaces import get_interface_attr
 from openpilot.selfdrive.car.fingerprints import eliminate_incompatible_cars, all_legacy_fingerprint_cars
@@ -206,14 +207,33 @@ def get_car(logcan, sendcan, experimental_long_allowed, num_pandas=1):
 
   return get_car_interface(CP), CP
 
+
 def write_car_param(platform=MOCK.MOCK):
   params = Params()
   CarInterface, _, _ = interfaces[platform]
   CP = CarInterface.get_non_essential_params(platform)
   params.put("CarParams", CP.to_bytes())
 
+
 def get_demo_car_params():
   platform = MOCK.MOCK
   CarInterface, _, _ = interfaces[platform]
   CP = CarInterface.get_non_essential_params(platform)
   return CP
+
+
+def apply_center_deadzone(error, deadzone):
+  if (error > - deadzone) and (error < deadzone):
+    error = 0.
+  return error
+
+
+def get_friction(lateral_accel_error: float, lateral_accel_deadzone: float, friction_threshold: float,
+                 torque_params: car.CarParams.LateralTorqueTuning, friction_compensation: bool) -> float:
+  friction_interp = interp(
+    apply_center_deadzone(lateral_accel_error, lateral_accel_deadzone),
+    [-friction_threshold, friction_threshold],
+    [-torque_params.friction, torque_params.friction]
+  )
+  friction = float(friction_interp) if friction_compensation else 0.0
+  return friction
